# movie-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "movie-api microservice",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/movies": {
            "get": {
                "tags": [
                    "Movie"
                ],
                "summary": "Affichage des films",
                "description": "Renvoie les films",
                "operationId": "a7859dc5cd106bbeae5492e22df893ec",
                "responses": {
                    "201": {
                        "description": "Affiche les films",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Movie": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Movie"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unauthorized"
                    }
                }
            },
            "put": {
                "tags": [
                    "Movie"
                ],
                "summary": "Modification d'un film",
                "description": "Renvoie un film",
                "operationId": "cb7850b94a42b497e938d19c0cfc6c1e",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "name": "UpdateMovieRequest",
                        "in": "query",
                        "required": true,
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UpdateMovieRequest"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Film modifié avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Movie"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Movie"
                ],
                "summary": "Création d'un film",
                "description": "Renvoie un film",
                "operationId": "a303e8fe5382b02a193fa994ebef739f",
                "parameters": [
                    {
                        "name": "CreateMovieRequest",
                        "in": "query",
                        "required": true,
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StoreMovieRequest"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Film crée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Movie"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unauthorized"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Movie"
                ],
                "summary": "Suppression d'un film",
                "description": "Ne renvoie rien",
                "operationId": "ae629d07b8324397c80287ba10dcb27a",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Film supprimé avec succès"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/movies/{movie}": {
            "get": {
                "tags": [
                    "Movie"
                ],
                "summary": "Affichage d'un film",
                "description": "Renvoie un film",
                "operationId": "36a46ca60c0936c9a2e4abd20099c5e5",
                "responses": {
                    "201": {
                        "description": "Affiche un film",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Movie"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unauthorized"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "StoreMovieRequest": {
                "title": "StoreMovieRequest",
                "description": "Store movie model request",
                "required": [
                    "name",
                    "description",
                    "released_at"
                ],
                "properties": {
                    "name": {
                        "description": "Movie name",
                        "type": "string"
                    },
                    "description": {
                        "description": "Movie description",
                        "type": "string"
                    },
                    "released_at": {
                        "description": "Movie release date",
                        "type": "string",
                        "format": "date"
                    },
                    "rate": {
                        "description": "Movie rate",
                        "type": "number",
                        "maximum": 5,
                        "minimum": 1
                    },
                    "duration": {
                        "description": "Movie duration",
                        "type": "number",
                        "maximum": 239,
                        "minimum": 1
                    }
                },
                "type": "object"
            },
...
            "UpdateMovieRequest": {
                "title": "UpdateMovieRequest",
                "description": "Update movie model request",
                "properties": {
                    "name": {
                        "description": "Movie name",
                        "type": "string"
                    },
                    "description": {
                        "description": "Movie description",
                        "type": "string"
                    },
                    "rate": {
                        "description": "Movie rate",
                        "type": "number",
                        "maximum": 5,
                        "minimum": 1
                    },
                    "duration": {
                        "description": "Movie duration",
                        "type": "number",
                        "maximum": 239,
                        "minimum": 1
                    }
                },
                "type": "object"
            },
...
            "Movie": {
                "title": "Movie",
                "description": "Movie model",
                "required": [
                    "id",
                    "name",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "Movie name",
                        "type": "string"
                    },
                    "description": {
                        "description": "Movie description",
                        "type": "string"
                    },
                    "released_at": {
                        "description": "Movie release date",
                        "type": "string",
                        "format": "date"
                    },
                    "rate": {
                        "description": "Movie rate",
                        "type": "number",
                        "maximum": 5,
                        "minimum": 1
                    },
                    "duration": {
                        "description": "Movie duration",
                        "type": "number",
                        "maximum": 239,
                        "minimum": 1
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-05T12:34:56.789Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-05T12:34:56.789Z"
                    }
                },
                "type": "object"
            }
        }
    }
}